Add patch from Eric Roccasecca (via email)
Guards against an issue when starting the spinner multiple times. It will
now always kill the timer and make a new one when told to start animating,
instead of starting more times (which cause the animation to run faster
and faster).

Also fixes a potential leak of the spinner object. The spinner object
would never get fully released if it was animating since timers retain
their targets. Does this by implementing -viewDidMoveToWindow to handle
the case where the window is nil and kills the timer in that case.