Added display when stopped
* Add new setting to YRKSpinningProgressIndicator.
* Add checkbox to demo app to demonstrate it.

Based on pull request from Michael Schneider, but I tweaked it a bit to put
the show/hide logic into YRKSpinningProgressIndicator instead of in the
demo app, fixed it to not hide it while it's animating, and fixed the
spelling of an argument.  I also omitted all the changes in the project file.